+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,7 +1,5 @@
-package racingcar.service;
+package racingcar.domain;
 
-import racingcar.domain.Car;
-import racingcar.domain.Cars;
 import java.util.List;
 import java.util.Map;
 import racingcar.strategy.MovingStrategy;
